Adicionado GET, PATCH, POST e DELETE

+++ src/main/java/br/com/estacionamento/estacionAqui/repository/ClienteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.estacionamento.estacionAqui.model.Cliente;
 
 @Repository   // classe para interagir com banco de dados
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
-
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+	 
 }
